Move EdgeConvTito to components/edge_convolutions.py
